Allow empty replies not to send a response
Now that we memoize the getEnhancedCupcake function, we have to be
deliberate about failures. That is, the previous behavior for
getEnhancedCupcake was to resolve the promise positively with a string
message describing the failure. However, with memoization, resolving the
promise means the result gets cached, which we don't want.

So, we need to throw if the cupcake doesn't work, which means we need to
catch replies. Since the cupcake error has use, we need to respond with
caught errors to Flowdock. Since errors are now sent to Flowdock, we
need to have a way for non-error, non-useful replies _not_ to get sent
to Flowdock.