(getset) Prevent .set from storing an empty file
When constructing a reply to a message, we do a number of checks that
the reply is truthy before returning it to the user. If a file is
empty, the `reply` is an empty string and is falsy, meaning that we
don't reply, which is confusing behavior.

So, if the user messes up the `.set` command or tries to store an
empty value, we instead prefer to put an `` in, which will evaluate
truthy.